Ejh valgrind 0329

+++ libdispatch/dfile.c
@@ -21,6 +21,7 @@
 #ifdef HAVE_SYS_STAT_H
 #include <sys/stat.h>
 #endif
+
 #ifdef HAVE_UNISTD_H
 #include <unistd.h> /* lseek() */
 #endif
@@ -181,7 +182,7 @@ NC_interpret_magic_number(char* magic, int* model, int* version)
         if(magic[3] == '\001') {
             *version = 1; /* netcdf classic version 1 */
 	    *model = NC_FORMATX_NC3;
-	    goto done;	    
+	    goto done;
 	}
         if(magic[3] == '\002') {
             *version = 2; /* netcdf classic version 2 */
@@ -234,7 +235,7 @@ NC_check_file_type(const char *path, int flags, void *parameters,
    *model = 0;
    *version = 0;
 
-    memset((void*)&file,0,sizeof(file));   
+    memset((void*)&file,0,sizeof(file));
     file.path = path; /* do not free */
     file.parameters = parameters;
     if(inmemory && parameters == NULL)
@@ -283,7 +284,7 @@ next:
 	    /* double and try again */
 	    pos = 2*pos;
         }
-    }    
+    }
 done:
     closemagic(&file);
     return status;
@@ -574,7 +575,7 @@ nc__create(const char *path, int cmode, size_t initialsz,
 }
 /**
  * @internal Create a file with special (deprecated) Cray settings.
- * 
+ *
  * @deprecated This function was used in the old days with the Cray at
  * NCAR. The Cray is long gone, and this call is supported only for
  * backward compatibility. Use nc_create() instead.
@@ -626,7 +627,7 @@ nc__create_mp(const char *path, int cmode, size_t initialsz,
  * to the dataset, including appending or changing data, adding or
  * renaming dimensions, variables, and attributes, or deleting
  * attributes.)
- * 
+ *
  * The NC_SHARE flag is only used for netCDF classic and 64-bit offset
  * files. It is appropriate when one process may be writing the
  * dataset and one or more other processes reading the dataset
@@ -665,14 +666,14 @@ nc__create_mp(const char *path, int cmode, size_t initialsz,
  * accessed data to be retrieved from disk. Without MMAP, NC_DISKLESS
  * will read the whole file into memory on nc_open. Thus, MMAP will
  * provide some performance improvement in this case.
- * 
+ *
  * It is not necessary to pass any information about the format of the
  * file being opened. The file type will be detected automatically by
  * the netCDF library.
  *
  * If a the path is a DAP URL, then the open mode is read-only.
  * Setting NC_WRITE will be ignored.
- * 
+ *
  * As of version 4.3.1.2, multiple calls to nc_open with the same
  * path will return the same ncid value.
  *
@@ -694,10 +695,10 @@ nc__create_mp(const char *path, int cmode, size_t initialsz,
  * @returns ::NC_EDIMMETA Error in netCDF-4 dimension metadata. (NetCDF-4 files only.)
  *
  * <h1>Examples</h1>
- * 
+ *
  * Here is an example using nc_open()to open an existing netCDF dataset
  * named foo.nc for read-only, non-shared access:
- * 
+ *
  * @code
  * #include <netcdf.h>
  *   ...
@@ -1722,12 +1723,12 @@ check_create_mode(int mode)
        (mode & NC_CDF5);
     if (mode_format && (mode_format & (mode_format - 1)))
        return NC_EINVAL;
-    
+
     /* Can't use both NC_MPIIO and NC_MPIPOSIX. Make up your damn
      * mind! */
     if (mode & NC_MPIIO && mode & NC_MPIPOSIX)
        return NC_EINVAL;
-    
+
     /* Can't use both parallel and diskless. */
     if ((mode & NC_MPIIO && mode & NC_DISKLESS) ||
 	(mode & NC_MPIPOSIX && mode & NC_DISKLESS))
@@ -1736,11 +1737,11 @@ check_create_mode(int mode)
 #ifndef USE_DISKLESS
    /* If diskless is requested, but not built, return error. */
    if (mode & NC_DISKLESS)
-       return NC_ENOTBUILT;       
+       return NC_ENOTBUILT;
    if (mode & NC_INMEMORY)
-       return NC_ENOTBUILT;       
+       return NC_ENOTBUILT;
 #endif
-   
+
 #ifndef USE_NETCDF4
    /* If the user asks for a netCDF-4 file, and the library was built
     * without netCDF-4, then return an error.*/
@@ -1811,7 +1812,7 @@ NC_create(const char *path0, int cmode, size_t initialsz,
    /* Check mode flag for sanity. */
    if ((stat = check_create_mode(cmode)))
       return stat;
-   
+
    /* Initialize the dispatch table. The function pointers in the
     * dispatch table will depend on how netCDF was built
     * (with/without netCDF-4, DAP, CDMREMOTE). */
@@ -1969,7 +1970,7 @@ NC_create(const char *path0, int cmode, size_t initialsz,
 
    if((cmode & NC_MPIIO) && (cmode & NC_MPIPOSIX))
    {
-       nullfree(path);       
+       nullfree(path);
        return  NC_EINVAL;
    }
 
@@ -2383,14 +2384,14 @@ NC_open(const char *path0, int cmode, int basepe, size_t *chunksizehintp,
          dispatcher = NC3_dispatch_table;
          break;
       default:
-         nullfree(path);              
+         nullfree(path);
          return NC_ENOTNC;
       }
    }
 
    /* If we can't figure out what dispatch table to use, give up. */
    if (!dispatcher) {
-       nullfree(path);              
+       nullfree(path);
        return NC_ENOTNC;
    }
 
@@ -2572,12 +2573,12 @@ readmagic(struct MagicFile* file, long pos, char* magic)
     }
 #endif /* USE_PARALLEL */
     {
-	size_t count;
+	int count;
 	int i = fseek(file->fp,pos,SEEK_SET);
 	if(i < 0)
 	    {status = errno; goto done;}
 	for(i=0;i<MAGIC_NUMBER_LEN;) {/* make sure to read proper # of bytes */
-	    count=fread(&magic[i],1,(MAGIC_NUMBER_LEN-i),file->fp);
+	    count=fread(&magic[i],1,(size_t)(MAGIC_NUMBER_LEN-i),file->fp);
 	    if(count == 0 || ferror(file->fp))
 		{status = errno; goto done;}
 	    i += count;
@@ -2589,9 +2590,9 @@ done:
     return status;
 }
 
-/** 
- * Close the file opened to check for magic number. 
- * 
+/**
+ * Close the file opened to check for magic number.
+ *
  * @param file pointer to the MagicFile struct for this open file.
  * @returns NC_NOERR for success
  * @returns NC_EPARINIT if there was a problem closing file with MPI
