+++ libsrc/nc3internal.c
@@ -711,7 +711,7 @@ NC_check_vlens(NC3_INFO *ncp)
     /* maximum permitted variable size (or size of one record's worth
        of a record variable) in bytes.  This is different for format 1
        and format 2. */
-    size_t vlen_max;
+    unsigned long long vlen_max;
     size_t ii;
     size_t large_vars_count;
     size_t rec_vars_count;
@@ -721,7 +721,7 @@ NC_check_vlens(NC3_INFO *ncp)
 	return NC_NOERR;
 
     if (fIsSet(ncp->flags,NC_64BIT_DATA)) /* CDF-5 */
-	vlen_max = X_INT64_MAX - 3; /* "- 3" handles rounded-up size */
+	vlen_max = (size_t)X_INT64_MAX - 3; /* "- 3" handles rounded-up size */
     else if (fIsSet(ncp->flags,NC_64BIT_OFFSET) && sizeof(off_t) > 4)
 	/* CDF2 format and LFS */
 	vlen_max = X_UINT_MAX - 3; /* "- 3" handles rounded-up size */
